requirements.txt
streamlit
pymupdf
python-docx
pandas
plotly
supabase
reportlab
python-dotenv